Moved AbstractProcessor out of core package
